Microweb additions and other fixes

Approved-by: Ken B <broken2048@yahoo.com>
Approved-by: Todd Giles <todd.giles@gmail.com>

+++ web/microweb/cgi-bin/web_controls.cgi
@@ -8,8 +8,9 @@
 
 import cgi
 import cgitb
-import os
+import os, stat
 import subprocess
+import ConfigParser
 from time import sleep
 
 
@@ -19,6 +20,31 @@ message = form.getvalue("message", "")
 
 HOME_DIR = os.getenv("SYNCHRONIZED_LIGHTS_HOME")
 volume = subprocess.check_output([HOME_DIR + '/bin/vol'])
+
+state_file = HOME_DIR + '/web/microweb/config/webstate.cfg'
+state = ConfigParser.RawConfigParser()
+state.readfp(open(state_file))
+config_file = state.get('microweb','config')
+if config_file:
+    config_param = '--config=' + config_file + ' '
+else:
+    config_param = ''
+    if os.path.isfile(HOME_DIR + '/config/overrides.cfg'):
+        config_file = 'overrides.cfg'
+    else:
+        config_file = 'defaults.cfg'
+
+cfg_file = HOME_DIR + '/config/' + config_file
+cfg = ConfigParser.RawConfigParser()
+cfg.readfp(open(cfg_file))
+lightshowmode = cfg.get('lightshow','mode')
+lightshowstc = cfg.get('lightshow','stream_command_string')
+
+if lightshowmode == "stream-in" and lightshowstc == "pianobar":
+    try:
+        os.stat('/root/.config/pianobar/ctl')
+    except OSError:
+        os.system('mkfifo /root/.config/pianobar/ctl')
 
 if message:
     if message == "Volume -":
@@ -36,19 +62,23 @@ if message:
     if message == "On":
         os.system('pkill -f "bash $SYNCHRONIZED_LIGHTS_HOME/bin"')
         os.system('pkill -f "python $SYNCHRONIZED_LIGHTS_HOME/py"')
-        os.system("python ${SYNCHRONIZED_LIGHTS_HOME}/py/hardware_controller.py --state=on")
+        os.system("python ${SYNCHRONIZED_LIGHTS_HOME}/py/hardware_controller.py " + config_param + "--state=on")
     if message == "Off":
         os.system('pkill -f "bash $SYNCHRONIZED_LIGHTS_HOME/bin"')
         os.system('pkill -f "python $SYNCHRONIZED_LIGHTS_HOME/py"')
-        os.system("python ${SYNCHRONIZED_LIGHTS_HOME}/py/hardware_controller.py --state=off")
-    if message == "Next":
+        os.system("python ${SYNCHRONIZED_LIGHTS_HOME}/py/hardware_controller.py " + config_param + "--state=off")
+        sleep(2)
+    if message == "Next" and lightshowmode == "playlist":
         os.system('pkill -f "python $SYNCHRONIZED_LIGHTS_HOME/py"')
+        sleep(1)
+    if message == "Next" and lightshowmode == "stream-in" and lightshowstc == "pianobar":
+        os.system('echo -n "n" > /root/.config/pianobar/ctl')
         sleep(1)
     if message == "Start":
         os.system('pkill -f "bash $SYNCHRONIZED_LIGHTS_HOME/bin"')
         os.system('pkill -f "python $SYNCHRONIZED_LIGHTS_HOME/py"')
-        os.system("${SYNCHRONIZED_LIGHTS_HOME}/bin/play_sms &")
-        os.system("${SYNCHRONIZED_LIGHTS_HOME}/bin/check_sms &")
+        os.system("${SYNCHRONIZED_LIGHTS_HOME}/bin/play_sms " + config_param + "&")
+        os.system("${SYNCHRONIZED_LIGHTS_HOME}/bin/check_sms " + config_param + "&")
         sleep(1)
 
 print "Content-type: text/html"
@@ -69,26 +99,40 @@ print """
         <link rel="apple-touch-icon" sizes="152x152" href="/favicon.png">
         <link rel="stylesheet" href="/css/style.css">
     </head>
-    <body>
-        <center>
+    <body class="centered-wrapper">
             <h2> LightShowPi Web Controls </h2>
+
+            <table class="centered-content">
+            <tr><td>
+            <form method="post" action="tools.cgi">
+                <input id="tools" type="image" src="/toolsicon64.png" >
+            </form>
+            </td><td>
+            <form method="post" action="settings.cgi">
+                <input id="settings" type="image" src="/gearicon64.png" >
+            </form>
+            </td></tr>
+            </table>
 
             <div id="voldiv">
             <form method="post" action="web_controls.cgi">
                 <input id="volDown" type="submit" name="message" value="Volume -">
 """
 
-print "&nbsp" + volume + "&nbsp"
+print '<div id="volumediv" class="centered-content">' + volume + '</div>'
 
 print """
                 <input id="volUp" type="submit" name="message" value="Volume +">
             </form>
             </div>
-
+"""
+if lightshowmode == "playlist":
+    print """
             <form method="post" action="playlist.cgi">
                 <input id="playlist" type="submit" value="Playlist">
             </form>
-
+"""
+print """
             <form method="post" action="web_controls.cgi">
                 <input type="hidden" name="message" value="On"/>
                 <input id="on" type="submit" value="Lights ON">
